+++ DG2D/dfr_startup.go
@@ -20,10 +20,11 @@ type DFR2D struct {
 	// FluxElement        *RTBasis2DSimplexLegacy
 	FluxElement *RTElement
 	// Interpolates from the interior (solution) points to graph points
-	GraphMesh          geometry.TriMesh
-	GraphInterp        utils.Matrix
-	FluxEdgeInterp     utils.Matrix // Interpolates only from interior to the edge points in the flux element
-	FilterMod          utils.Matrix // Solution Modulation filter damps high modes
+	GraphMesh      geometry.TriMesh
+	GraphInterp    utils.Matrix
+	GraphInterpMod utils.Matrix // Modulated
+	FluxEdgeInterp utils.Matrix // Interpolates only from interior to the edge points in the flux element
+	// FilterMod          utils.Matrix // Solution Modulation filter damps high modes
 	FluxDr, FluxDs     utils.Matrix // Derivatives from the interior (solution) points to all of the flux points
 	DXMetric, DYMetric utils.Matrix // X and Y derivative metrics, multiply by scalar field values to create DOF for RT
 	// Mesh Parameters
@@ -61,11 +62,14 @@ func NewDFR2D(N int, verbose bool, meshFileO ...string) (dfr *DFR2D) {
 	// Get the interpolation matrices that interpolate the whole RT element and just the edges using solution points
 	GraphR, GraphS := dfr.GetRSForGraphMesh()
 	Nu, p := 0.1, 3.29
-	dfr.GraphInterp = dfr.SolutionBasis.GetModInterpMatrix(GraphR, GraphS, Nu, p, 1)
-	dfr.FluxEdgeInterp = dfr.SolutionBasis.GetInterpMatrix(RFlux, SFlux) // Interpolation matrix across three edges
+	// dfr.GraphInterp = dfr.SolutionBasis.GetModInterpMatrix(GraphR, GraphS, Nu, p, 1)
+	dfr.GraphInterpMod = dfr.SolutionBasis.GetModInterpMatrix(GraphR, GraphS,
+		Nu, p, 1)
+	dfr.GraphInterp = dfr.SolutionBasis.GetInterpMatrix(GraphR, GraphS)
+	// dfr.FluxEdgeInterp = dfr.SolutionBasis.GetInterpMatrix(RFlux, SFlux) // Interpolation matrix across three edges
+	dfr.FluxEdgeInterp = dfr.SolutionBasis.GetModInterpMatrix(RFlux, SFlux, Nu, p, 1) // Interpolation matrix across three edges
 	// dfr.FluxEdgeInterp = dfr.SolutionBasis.GetModInterpMatrix(RFlux, SFlux, Nu, p, 1) // Interpolation matrix across three edges
-	// dfr.FluxEdgeInterp = dfr.SolutionBasis.GetModInterpMatrix(RFlux, SFlux, Nu, p, 1) // Interpolation matrix across three edges
-	dfr.FilterMod = dfr.SolutionBasis.GetModInterpMatrix(dfr.SolutionElement.R, dfr.SolutionElement.S, Nu, p, 1)
+	// dfr.FilterMod = dfr.SolutionBasis.GetModInterpMatrix(dfr.SolutionElement.R, dfr.SolutionElement.S, Nu, p, 1)
 	dfr.FluxDr, dfr.FluxDs = le.GetDerivativeMatrices(rt.R, rt.S)
 	if len(meshFileO) != 0 {
 		var EToV utils.Matrix